projects/due-diligence-agent/requirements.txt
# Core LLM Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0           # Optional: OpenAI support
langchain-ollama>=0.2.0           # Optional: Local LLM support

# Structured Output
pydantic>=2.0.0

# Search Tools
tavily-python>=0.5.0              # Primary search (free tier: 1000/month)
duckduckgo-search>=6.0.0          # Fallback search (no API key needed)

# Web & API
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0

# UI
streamlit>=1.40.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0                     # For FastAPI test client

# Utilities
tabulate>=0.9.0                   # Pretty table formatting
